ref(request): refactor request handling;
