+++ src/httpStatusCodes.ts
@@ -2,5 +2,7 @@ export const httpStatusCodes = {
   OK: 200,
   BAD_REQUEST: 400,
   NOT_FOUND: 404,
+  VALIDATION_FAILED: 404,
   INTERNAL_SERVER: 500,
+  SERVICE_UNAVAILABLE: 503,
 };
